Switched from being a gretina archive to a midas archiver.
The big change was that it becomes harder to tell when a run is done.
Gretina makes the folder for run N+1 when N ends, so you can notice it.
Midas only makes an output when the new run starts.
Therefore, waiting until a file has been unmodified for 3 minutes before copying.
As a safety check, continuing to watch it afterward as well.